implemented server iterator to retrieve data to dataRepository

+++ app/src/main/java/com/gustavo/bankandroid/entity/UserStatementItem.kt
@@ -5,5 +5,5 @@ data class UserStatementItem (
 	val title : String,
 	val desc : String,
 	val date : String,
-	val value : Int
+	val value : Float
 )
